fix(tests): Extract `assert_eval` to a separate `test_evaluation`.
- Eval now runs on all the traces.
- Running the eval as part of `test_agent` was causing timeout failures.
- Change `mistral-small` to `gpt-4.1-nano`. Mistral was just causing tiemouts even in separate tests.
