+++ tests/integration/test_agent.py
@@ -4,7 +4,6 @@ import subprocess
 import time
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import Any
 
 import pytest
 from litellm.utils import validate_environment
@@ -17,9 +16,6 @@ from any_agent import (
 )
 from any_agent.callbacks.span_print import ConsolePrintSpan
 from any_agent.config import MCPStdio
-from any_agent.evaluation.agent_judge import AgentJudge
-from any_agent.evaluation.llm_judge import LlmJudge
-from any_agent.evaluation.schemas import EvaluationOutput
 from any_agent.testing.helpers import (
     DEFAULT_SMALL_MODEL_ID,
     get_default_agent_model_args,
@@ -100,53 +96,6 @@ def assert_tokens(agent_trace: AgentTrace) -> None:
     assert agent_trace.tokens.input_tokens > 0
     assert agent_trace.tokens.output_tokens > 0
     assert (agent_trace.tokens.input_tokens + agent_trace.tokens.output_tokens) < 20000
-
-
-def assert_eval(agent_trace: AgentTrace) -> None:
-    """Test evaluation using the new judge classes."""
-    # Test 1: Check if agent called write_file tool using LlmJudge
-    llm_judge = LlmJudge(
-        model_id=DEFAULT_SMALL_MODEL_ID,
-        model_args={
-            "temperature": 0.0,
-        },  # Because it's an llm not agent, the default_model_args are not used
-    )
-    result1 = llm_judge.run(
-        context=str(agent_trace.spans_to_messages()),
-        question="Do the messages contain the year 2025?",
-    )
-    assert isinstance(result1, EvaluationOutput)
-    assert result1.passed, (
-        f"Expected agent to call write_file tool, but evaluation failed: {result1.reasoning}"
-    )
-
-    # Test 2: Check if agent wrote the current year to file using AgentJudge
-    agent_judge = AgentJudge(
-        model_id=DEFAULT_SMALL_MODEL_ID,
-        model_args=get_default_agent_model_args(AgentFramework.TINYAGENT),
-    )
-
-    def get_current_year() -> str:
-        """Get the current year"""
-        return str(datetime.now().year)
-
-    eval_trace = agent_judge.run(
-        trace=agent_trace,
-        question="Did the agent write the year to a file? Grab the messages from the trace and check if the write_file tool was called.",
-        additional_tools=[get_current_year],
-    )
-    result2 = eval_trace.final_output
-    assert isinstance(result2, EvaluationOutput)
-    assert result2.passed, (
-        f"Expected agent to write current year to file, but evaluation failed: {result2.reasoning}"
-    )
-
-    # Test 3: Verify at least one evaluation passes (basic sanity check)
-    results = [result1, result2]
-    passed_count = sum(1 for r in results if r.passed)
-    assert passed_count >= 1, (
-        f"Expected at least 1 evaluation to pass, but got {passed_count}/2"
-    )
 
 
 class Step(BaseModel):
@@ -256,9 +205,3 @@ def test_load_and_run_agent(
         html_output = console.export_html(inline_styles=True)
         with open(f"{trace_path}_trace.html", "w", encoding="utf-8") as f:
             f.write(html_output.replace("<!DOCTYPE html>", ""))
-
-    if (
-        agent_framework is AgentFramework.TINYAGENT
-        and model_id == DEFAULT_SMALL_MODEL_ID
-    ):
-        assert_eval(agent_trace)
